refactor: 디렉토리 개편
