+++ src/components/book/BookItem.css.ts
