+++ src/components/book/DetailSearch.tsx
@@ -3,7 +3,7 @@ import { Button, Popover, Select, Theme } from '@radix-ui/themes';
 import { type BookSearchTarget, bookSearchTargets } from 'src/types';
 import * as styles from './DetailSearch.css';
 import { X } from 'lucide-react';
-import { targetName } from 'src/lib/constant';
+import { targetName } from 'src/constants';
 
 interface DetailSearchProps {
   onSubmit: (query: string, target: 'title' | 'person' | 'publisher') => void;
