+++ src/components/index.ts
@@ -1,7 +1,2 @@
-export * from './shared';
-export { default as BookContext } from './BookContext';
-export { default as BookItem } from './BookItem';
-export { default as BookList } from './BookList';
-export { default as DetailSearch } from './DetailSearch';
-export { default as SearchBar } from './SearchBar';
-export { default as SearchHistory } from './SearchHistory';
+export * from './book';
+export * from './common';
